response-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
openai==0.28.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
mypy==1.7.1
tenacity==8.2.3
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-jaeger==1.22.0
python-json-logger==2.0.7
